Update types.ts
These attributes need to be optional. Otherwise if you create an EVM-specific RelayChainConfig you get error that it is missing 'pallets' attribute even though it shouldn't be required. Or if you create a Substrate-specific RelayChainConfig you get error that it is missing 'contracts' and 'relayerFeeConfig' even though they aren't relevant for Substrate.